Rewrite name handling and finish default instances (at least for now).
There's a few minor refactors in this commit as well, like finally
making builtin_types return things that don't need unwrapping with get,
but the title line is the soul of the changes.

Name handling is now done top-down, not bottom-up; every layer knows its
_full_ name as soon as it is possible. This simplified scope handling,
removed the need for scope generators, and enabled default instances.
Generally a lot of things are now how they should've been from the
start.

Anyway, it's now possible to import a default instance. Nice. Only...
a... year? in the works.